// api/generated/model/createMyRecipeBodySectionsItemIngredientsItemUnit.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * A Cheffy API é o backend central de uma aplicação de receitas, projetado para gerenciar catálogo culinário, categorias, tags, ingredientes, informações nutricionais, favoritos de usuário, autenticação via Better Auth, assinatura de upload no Cloudinary e um assistente gastronômico integrado à Groq.
 * OpenAPI spec version: 1.0.0
 */

export type CreateMyRecipeBodySectionsItemIngredientsItemUnit = typeof CreateMyRecipeBodySectionsItemIngredientsItemUnit[keyof typeof CreateMyRecipeBodySectionsItemIngredientsItemUnit];


export const CreateMyRecipeBodySectionsItemIngredientsItemUnit = {
  G: 'G',
  KG: 'KG',
  MG: 'MG',
  OZ: 'OZ',
  LB: 'LB',
  ML: 'ML',
  L: 'L',
  CUP: 'CUP',
  TBSP: 'TBSP',
  TSP: 'TSP',
  UNIT: 'UNIT',
  DOZEN: 'DOZEN',
  PINCH: 'PINCH',
  DRIZZLE: 'DRIZZLE',
  CUBE: 'CUBE',
  PACKAGE: 'PACKAGE',
  CAN: 'CAN',
  BOTTLE: 'BOTTLE',
  BOX: 'BOX',
  TO_TASTE: 'TO_TASTE',
  CLOVE: 'CLOVE',
  BUNCH: 'BUNCH',
  GLASS: 'GLASS',
  LEAF: 'LEAF',
  POT: 'POT',
} as const;